+++ tests/hypothesis_strategies/metadata.py
@@ -6,7 +6,7 @@ from hypothesis import assume
 from baybe.parameters.base import ParameterMetadata
 from baybe.utils.metadata import Metadata
 
-_descriptions = st.one_of(st.none(), st.text(min_size=1))
+_descriptions = st.one_of(st.none(), st.text(min_size=0))
 """A strategy generating metadata descriptions."""